Never consider TEST RUN locked
- Repairs access to TEST RUN cup
- We want to make it an unlockable later, so...
- M_MapLocked data types corrected

+++ src/m_cond.h
@@ -341,7 +341,7 @@ UINT8 M_CompletionEmblems(void);
 boolean M_CheckNetUnlockByID(UINT8 unlockid);
 boolean M_SecretUnlocked(INT32 type, boolean local);
 boolean M_CupLocked(cupheader_t *cup);
-boolean M_MapLocked(INT32 mapnum);
+boolean M_MapLocked(UINT16 mapnum);
 INT32 M_CountMedals(boolean all, boolean extraonly);
 
 // Emblem shit
